Let the volume show at the display box for two seconds when clicked
When you press a drumpad the name of it should show immediately on the display box. There is a bug here is that when
you press the same drumpad its name will not show immediately because useffect will not work. It only works when
"content" changes but content in this case doesn't change.